refactor: ts migration for poi/limited
