+++ packages/account/globals.d.ts
@@ -1,0 +1,7 @@
+export {};
+
+declare global {
+    interface Window {
+        LC_API: any;
+    }
+}
