+++ packages/components/stories/icon/icons.js
@@ -882,4 +882,4 @@ export const icons =
         'IcWalletZingpayDark',
         'IcWalletZingpayLight',
     ],
-};
+}
